Add new Renderer type to include full node data without changing node set
The nodes returned from Map2Parent are mostly empty, and need joining with their full topologies
before being useful. But we want to filter out any nodes that weren't already in the graph.

We make a new Renderer type to implement this.
We also add the missing types to the type option group